refactor: extract and modularize use-blockchains composable
